+++ security-spring/src/main/java/net/cofcool/chaos/server/security/spring/authorization/AbstractAuthenticationHandler.java
@@ -4,6 +4,9 @@ import net.cofcool.chaos.server.common.core.ExceptionCodeManager;
 import org.springframework.http.converter.HttpMessageConverter;
 import org.springframework.http.converter.json.MappingJackson2HttpMessageConverter;
 
+/**
+ * 针对 <code>AuthenticationHandler</code> 的基本配置项
+ */
 public abstract class AbstractAuthenticationHandler {
 
     private final ExceptionCodeManager exceptionCodeManager;