+++ CHANGELOG.md
@@ -31,3 +31,4 @@
 * `LoggingInterceptor`移除对`Jackson`的依赖, 改为调用对象的`toString`方法
 * 接口类的`getXxx()`风格代码改为`xxx()`
 * `UserAuthorizationService` 删除 `setUserProcessor` 方法
+* `security-spring`模块基本完成